refactor: change crypto_fee, fiat_fee columns type to DECIMAL

+++ prisma/migrations/20250329012259_add_crypto_fee_and_fiat_fee_columns_to_transaction/migration.sql
@@ -6,6 +6,6 @@
 */
 -- AlterTable
 ALTER TABLE "transaction" DROP COLUMN "gas_Used",
-ADD COLUMN     "crypto_fee" BIGINT,
-ADD COLUMN     "fiat_fee" INTEGER,
+ADD COLUMN     "crypto_fee" DECIMAL(65,30),
+ADD COLUMN     "fiat_fee" DECIMAL(65,30),
 ADD COLUMN     "gas_used" BIGINT;
